requirements.txt
Django==3.1
djangorestframework=3.12.4
djangorestframework-simplejwt==4.6.0
dnspython==1.16.0
httplib2==0.15.0
requests==2.22.0
Pillow==8.1.0
python-dotenv==0.19.0
psycopg2==2.9.1
django-cors-headers==3.6.0
gunicorn==20.1.0
